Bug fixes (#107)
* Stop arrow keys navigating move list when in a textarea

* Fix bug Play Computer mode crashing when another game ends in multiboard mode

* Fix bug game ends while movelist is being requested, movelistRequested not cleaned up properly

+++ src/index.ts
@@ -54,7 +54,6 @@ let historyRequested = 0;
 let obsRequested = 0;
 let allobsRequested = 0;
 let gamesRequested = false;
-let movelistRequested = 0;
 let lobbyRequested = false;
 let channelListRequested = false;
 let computerListRequested = false;
@@ -87,15 +86,12 @@ let partnerGameId = null;
 const mainBoard: any = createBoard($('#main-board-area').children().first().find('.board'));
 
 function cleanupGame(game: Game) {
-  if(game.role === Role.PLAYING_COMPUTER)
-    return;
-
-  game.role = Role.NONE;
-
-  if(playEngine) {
+  if(playEngine && game.role === Role.PLAYING_COMPUTER) {
     playEngine.terminate();
     playEngine = null;
   }
+
+  game.role = Role.NONE;
 
   if(game === gameWithFocus) {
     hideButton($('#stop-observing'));
@@ -129,6 +125,7 @@ function cleanupGame(game: Game) {
   gameExitPending = 0;
   game.partnerGameId = null;
   game.commitingMovelist = false;
+  game.movelistRequested = 0;
   game.board.cancelMove();
   updateBoard(game);
   initStatusPanel();
@@ -148,7 +145,6 @@ export function cleanup() {
   obsRequested = 0;
   allobsRequested = 0;
   gamesRequested = false;
-  movelistRequested = 0;
   lobbyRequested = false;
   channelListRequested = false;
   computerListRequested = false;
@@ -156,7 +152,8 @@ export function cleanup() {
   clearMatchRequests();
   clearNotifications();
   games.forEach((game) => {
-    cleanupGame(game);
+    if(game.role !== Role.PLAYING_COMPUTER)
+      cleanupGame(game);
   });
 }
 
@@ -1740,7 +1737,7 @@ function gameStart(game: Game) {
     }
 
     if(game.isExamining() || ((game.isObserving() || game.isPlayingOnline()) && game.move !== 'none')) {        
-      movelistRequested++;
+      game.movelistRequested++;
       session.send('iset startpos 1'); // Show the initial board position before the moves list 
       session.send('moves ' + game.id);
       session.send('iset startpos 0');
@@ -1840,10 +1837,8 @@ function messageHandler(data) {
             break;
           }
         }
-        else if(game.role === Role.PLAYING_COMPUTER && data.role !== Role.PLAYING_COMPUTER) {
-          game.role = Role.NONE;
+        else if(game.role === Role.PLAYING_COMPUTER && data.role !== Role.PLAYING_COMPUTER) 
           cleanupGame(game); // Allow player to imemediately play/examine/observe a game at any time while playing the Computer. The Computer game will simply be aborted.
-        }
       }
 
       if(examineModeRequested && data.role === Role.EXAMINING) {
@@ -1936,7 +1931,6 @@ function messageHandler(data) {
         }
         showBoardModal('Match Result', '', data.message, rematch, analyze, false, false, false);
       }
-      game.role = Role.NONE;
       cleanupGame(game);
       break;
     case MessageType.GameHoldings:
@@ -2335,10 +2329,8 @@ function messageHandler(data) {
       match = msg.match(/(?:^|\n)\s*Movelist for game (\d+):\s+(\w+) \((\d+|UNR)\) vs\. (\w+) \((\d+|UNR)\)[^\n]+\s+(\w+) (\S+) match, initial time: (\d+) minutes, increment: (\d+) seconds\./);
       if (match != null && match.length > 9) {
         var game = findGame(+match[1]);
-        if (movelistRequested) {
-          movelistRequested--;
-          if(!game)
-            return;
+        if(game && game.movelistRequested) {
+          game.movelistRequested--;
 
           if(game.isExamining()) {
             var id = match[1];
@@ -2502,8 +2494,12 @@ function messageHandler(data) {
 
       // Suppress messages when 'moves' command issued internally
       match = msg.match(/^You're at the (?:beginning|end) of the game\./m);
-      if(match && movelistRequested)
-        return;
+      if(match) {
+        for(let i = 0; i < games.length; i++) {
+          if(games[i].movelistRequested)
+            return;
+        }
+      }
 
       // Moving backwards and forwards is now handled more generally by updateHistory()
       match = msg.match(/^Game\s\d+: \w+ backs up (\d+) moves?\./m);
@@ -2791,7 +2787,7 @@ function updateHistory(game: Game, move?: any, fen?: string) {
   const hEntry = game.history.find(fen);
 
   if(!hEntry) {
-    if(movelistRequested)
+    if(game.movelistRequested)
       return;
 
     if(move) {
@@ -2813,14 +2809,14 @@ function updateHistory(game: Game, move?: any, fen?: string) {
     }
     else { 
       // move not found, request move list
-      movelistRequested++;
+      game.movelistRequested++;
       session.send('iset startpos 1'); // Show the initial board position before the moves list 
       session.send('moves ' + game.id);
       session.send('iset startpos 0');
     }
   }
   else {
-    if(!movelistRequested && game.role !== Role.NONE) 
+    if(!game.movelistRequested && game.role !== Role.NONE) 
       game.history.updateClockTimes(hEntry, game.wtime, game.btime);
 
     // move is already displayed
@@ -3098,7 +3094,7 @@ function startEngine() {
       options['UCI_Variant'] = game.category;
     
     engine = new Engine(game, null, displayEnginePV, options);
-    if(!movelistRequested)
+    if(!game.movelistRequested)
       engine.move(game.history.current());
   }
 }
@@ -3144,7 +3140,6 @@ function getPlayComputerMoveParams(game: Game): string {
 function playComputer(params: any) {
   var computerGame = getComputerGame();
   if(computerGame) {
-    computerGame.role = Role.NONE;
     cleanupGame(computerGame);
     var game = computerGame; 
   }
@@ -4934,7 +4929,7 @@ $('#exit-subvariation').on('click', () => {
 });
 
 $(document).on('keydown', (e) => {
-  if ($(e.target).closest('input')[0]) {
+  if ($(e.target).closest('input, textarea')[0]) {
     return;
   }
